requirements.txt
kubernetes
fastapi
uvicorn[standard]
gunicorn
pydantic[email]
passlib[bcrypt]
pyjwt
alembic
python-slugify
Unidecode
python-dotenv
databases
motor